transformers-use-cases/translation-bert-tr-en-finetuning-vertex/autopackage/requirements.txt
datasets==1.18.2
transformers==4.16.2
sentencepiece
sacrebleu==2.0.0
torch
google-cloud-storage